+++ Tools/GPUInfo/gpuinfo.c
@@ -2,11 +2,14 @@
 // GPU Info
 // Based on freevram and MetalInfo tools.
 // See: https://stackoverflow.com/questions/3783030/free-vram-on-os-x
-// 
+//
+
+#define OBJC_OLD_DISPATCH_PROTOTYPES 1
 
 #include <stdio.h>
 #include <stdint.h>
 #include <dlfcn.h>
+#include <unistd.h>
 
 #include <objc/runtime.h>
 #include <objc/message.h>
@@ -15,7 +18,7 @@
 #include <CoreFoundation/CoreFoundation.h>
 
 uint64_t currentFreeVRAM(uint64_t *total) {
-    kern_return_t krc;  
+    kern_return_t krc;
     mach_port_t masterPort;
     krc = IOMasterPort(bootstrap_port, &masterPort);
 
@@ -56,13 +59,13 @@ uint64_t currentFreeVRAM(uint64_t *total) {
     return 0;
 }
 
-int main() {
+void printInfo(void) {
     uint64_t total_vram = 0;
     uint64_t free_vram = currentFreeVRAM(&total_vram);
     if (total_vram > 0)
-        printf("Total VRAM availabile: %zu MB\n", (size_t)total_vram);
+        printf("Total VRAM available: %zu MB\n", (size_t)total_vram);
     if (free_vram > 0)
-        printf("Free VRAM availabile: %zu MB (%zu Bytes)\n", (size_t)(free_vram/(1024*1024)), (size_t)free_vram);
+        printf("Free VRAM available: %zu MB (%zu Bytes)\n", (size_t)(free_vram/(1024*1024)), (size_t)free_vram);
 
     void *mtl = dlopen("/System/Library/Frameworks/Metal.framework/Metal", RTLD_LAZY);
     id (*create_dev)(void) = (id (*)(void))dlsym(mtl, "MTLCreateSystemDefaultDevice");
@@ -80,5 +83,17 @@ int main() {
         }
     } else {
         printf("Metal is not supported by this operating system!\n");
+    }
+}
+
+int main(int argc, char *argv[]) {
+    if (argc > 1 && !strcmp(argv[1], "-l")) {
+        while (1) {
+            system("clear");
+            printInfo();
+            sleep(1);
+        }
+    } else {
+        printInfo();
     }
 }