+++ WhateverGreen/kern_ngfx.hpp
@@ -13,6 +13,7 @@
 #include <Headers/kern_patcher.hpp>
 #include <Headers/kern_devinfo.hpp>
 #include <IOKit/IOService.h>
+#include <IOKit/ndrvsupport/IONDRVFramebuffer.h>
 
 // Assembly exports for restoreLegacyOptimisations
 extern "C" bool wrapVaddrPreSubmitTrampoline(void *that);
@@ -66,6 +67,11 @@ private:
 	int fifoSubmit {-1};
 
 	/**
+	 *  Enable debug logging in NVIDIA drivers
+	 */
+	bool enableDebugLogging {false};
+
+	/**
 	 *  Disable team unrestriction patches fixing visual glitches on 10.12 with Web drivers
 	 */
 	bool disableTeamUnrestrict {false};
@@ -105,6 +111,11 @@ private:
 	 *  Original NVDAStartupWeb::probe function
 	 */
 	mach_vm_address_t orgStartupWebProbe {};
+
+	/**
+	 *  Original IONDRVFramebuffer::_doControl function
+	 */
+	mach_vm_address_t orgNdrvDoControl {};
 
 	/**
 	 *  Restore legacy optimisations from 10.13.0, which fix lags for Kepler GPUs.
@@ -152,6 +163,16 @@ private:
 	 *  NVDAStartup::probe wrapper used to force-enable web-drivers
 	 */
 	static IOService *wrapStartupWebProbe(IOService *that, IOService *provider, SInt32 *score);
+
+	/**
+	 *  IONDRVFramebuffer::_doControl wrapper used to avoid debug spam
+	 */
+	static IOReturn wrapNdrvDoControl(IONDRVFramebuffer *fb, UInt32 code, void *params);
+
+	/**
+	 *  nvErrorLog_va replacement with logging support
+	 */
+	static void resmanErrorLogVA(void *context, uint32_t id, const char *format, ...);
 };
 
 #endif /* kern_ngfx_hpp */